+++ core/lib.rs
@@ -20,6 +20,7 @@ mod resources;
 mod runtime;
 pub mod snapshot_util;
 mod source_map;
+mod task_queue;
 
 // Re-exports
 pub use anyhow;
@@ -116,6 +117,7 @@ pub use crate::runtime::Snapshot;
 pub use crate::runtime::V8_WRAPPER_OBJECT_INDEX;
 pub use crate::runtime::V8_WRAPPER_TYPE_INDEX;
 pub use crate::source_map::SourceMapGetter;
+pub use crate::task_queue::TaskQueue;
 
 pub fn v8_version() -> &'static str {
   v8::V8::get_version()
